refactor: organize remaining route tests by domain
