+++ backend/tests/routes/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Auth route tests"""
