+++ backend/tests/routes/cache/__init__.py
@@ -1,0 +1,1 @@
+"""Cache route tests"""
