+++ backend/tests/routes/cache/test_status.py
@@ -1,6 +1,6 @@
 from datetime import UTC, datetime
 
-from tests.routes.conftest import _create_account, _create_user, _get_auth_header
+from tests.routes.support import _create_account, _create_user, _get_auth_header
 
 
 async def _create_second_user(client):
