+++ backend/tests/routes/currencies/__init__.py
@@ -1,0 +1,1 @@
+"""Currency route tests"""
