+++ backend/tests/routes/groups/__init__.py
@@ -1,0 +1,1 @@
+"""Group route tests"""
