+++ backend/tests/routes/insights/__init__.py
@@ -1,0 +1,1 @@
+"""Insight route tests"""
