+++ backend/tests/routes/insights/test_net_worth.py
@@ -9,7 +9,7 @@ from sqlalchemy import update
 from app.models.account import AccountBalanceSnapshot
 from app.models.currency import Currency
 from tests.conftest import TestSession
-from tests.routes.conftest import _create_account, _create_user, _get_auth_header
+from tests.routes.support import _create_account, _create_user, _get_auth_header
 
 
 def _snapshot(account_id: UUID, dt: date, balance: int) -> AccountBalanceSnapshot:
