+++ backend/tests/routes/institutions/__init__.py
@@ -1,0 +1,1 @@
+"""Institution route tests"""
