+++ backend/tests/routes/institutions/test_institutions.py
@@ -1,7 +1,7 @@
 from app.models.base import InstitutionStatus
 from app.models.institution import Institution
 from tests.conftest import TestSession
-from tests.routes.conftest import _create_user, _get_auth_header
+from tests.routes.support import _create_user, _get_auth_header
 
 # --- Helpers ---
