+++ backend/tests/routes/tax_advantaged_categories/__init__.py
@@ -1,0 +1,1 @@
+"""Tax-advantaged category route tests"""
