+++ backend/tests/routes/users/__init__.py
@@ -1,0 +1,1 @@
+"""User route tests"""
